requirements.txt
attrs==24.2.0
awscli==1.34.8
botocore==1.35.8
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
docutils==0.16
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.8
imageio==2.35.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.4
pyasn1==0.6.0
PyMatting==1.1.12
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
sympy==1.13.2
tifffile==2024.8.28
tqdm==4.66.5
urllib3==1.26.19
